chore: remove unused API resources and handlers (#3641)
Remove 5 unused API endpoints that had TypeScript wrappers but were
never called from production code: POST /api/entrypoints,
DELETE /api/configurations/{name}, DELETE /api/deployments/{name},
GET /api/credentials/{guid}, and GET /api/configurations/{name}/secrets.
