+++ extensions/vscode/src/api/index.ts
@@ -10,5 +10,4 @@ export * from "./types/contentRecords";
 export * from "./types/events";
 export * from "./types/files";
 export * from "./types/schema";
-export * from "./types/entrypoints";
 export * from "./types/snowflakeConnections";
