+++ extensions/vscode/src/api/resources/Credentials.ts
@@ -71,14 +71,6 @@ export class Credentials {
   }
 
   // Returns:
-  // 200 - accepted
-  // 404 - not found
-  // 500 - internal server error
-  get(guid: string) {
-    return this.client.get<Credential>(`credentials/${guid}`);
-  }
-
-  // Returns:
   // 204 - no content
   // 404 - not found
   // 500 - internal server error
