--- extensions/vscode/src/api/types/entrypoints.ts
@@ -1,3 +1,0 @@
-// Copyright (C) 2023 by Posit Software, PBC.
-
-export type EntryPointPath = string;
