+++ extensions/vscode/src/constants.ts
@@ -12,7 +12,6 @@ export const DEPLOYMENTS_PATTERN = "**/.posit/publish/deployments/*.toml";
 export const DEFAULT_PYTHON_PACKAGE_FILE = "requirements.txt";
 export const DEFAULT_R_PACKAGE_FILE = "renv.lock";
 
-// pulled from /internal/services/api/get_entrypoints.go
 // should all be lowercase!
 export const ENTRYPOINT_FILE_EXTENSIONS = [
   ".htm",
